+++ src/V1/Internal/IsDuplicateContent.php
@@ -34,43 +34,46 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   LocalFilesystem\V1\TypeConverters
+ * @package   LocalFilesystem\V1\Internal
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://ganbarodigital.github.io/php-mv-filesystem-plugin-local
  */
 
-namespace GanbaroDigital\LocalFilesystem\V1\TypeConverters;
+namespace GanbaroDigital\LocalFilesystem\V1\Internal;
 
-use GanbaroDigital\AdaptersAndPlugins\V1\PluginTypes\PluginClass;
-use GanbaroDigital\LocalFilesystem\V1\Internal;
-use GanbaroDigital\LocalFilesystem\V1\LocalFileInfo;
-use GanbaroDigital\LocalFilesystem\V1\LocalFilesystem;
-use GanbaroDigital\MissingBits\ErrorResponders\OnFatal;
-
-class ToFileInfo implements PluginClass
+/**
+ * check for duplicate content
+ */
+class IsDuplicateContent
 {
     /**
-     * convert a path into a LocalFileInfo value
+     * check for duplicate content
      *
-     * @param  LocalFilesystem $fs
-     *         the filesystem that $path lives on
-     * @param  string|LocalFileInfo $path
-     *         the path that we want to convert
-     * @param  OnFatal $onFatal
-     *         what do we do if we cannot do the type conversion?
-     * @return LocalFileInfo
+     * @param  string $fullPath
+     *         path to the file we want to check
+     * @param  string $newContent
+     *         the new content that we want to check against
+     * @return bool
+     *         - TRUE if there is an existing file, and it contains
+     *           exactly what we are about to write
+     *         - FALSE otherwise
      */
-    public static function from(LocalFilesystem $fs, $path, OnFatal $onFatal) : LocalFileInfo
+    public static function check(string $fullPath, string $newContent) : bool
     {
-        // does it need converting?
-        if ($path instanceof LocalFileInfo) {
-            // no, it does not
-            return $path;
+        // does the file exist?
+        if (!is_file($fullPath)) {
+            return false;
         }
 
-        // it needs converting
-        return Internal\GetFileInfo::for($fs, $path, $onFatal);
+        $existingContents = file_get_contents($fullPath);
+
+        if (strcmp($existingContents, $newContent) !== 0) {
+            return false;
+        }
+
+        // if we get here, the content is equal
+        return true;
     }
 }