+++ src/V1/Operations/GetFileMetadata.php
@@ -33,44 +33,51 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * @category  Libraries
- * @package   LocalFilesystem\V1\TypeConverters
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://ganbarodigital.github.io/php-mv-filesystem-plugin-local
  */
 
-namespace GanbaroDigital\LocalFilesystem\V1\TypeConverters;
+namespace GanbaroDigital\LocalFilesystem\V1\Operations;
 
 use GanbaroDigital\AdaptersAndPlugins\V1\PluginTypes\PluginClass;
+use GanbaroDigital\Filesystem\V1\Checks;
+use GanbaroDigital\Filesystem\V1\TypeConverters;
 use GanbaroDigital\LocalFilesystem\V1\Internal;
-use GanbaroDigital\LocalFilesystem\V1\LocalFileInfo;
 use GanbaroDigital\LocalFilesystem\V1\LocalFilesystem;
 use GanbaroDigital\MissingBits\ErrorResponders\OnFatal;
 
-class ToFileInfo implements PluginClass
+/**
+ * return the metadata attached to a file
+ */
+class GetFileMetadata implements PluginClass
 {
     /**
-     * convert a path into a LocalFileInfo value
+     * return the contents of a file
      *
-     * @param  LocalFilesystem $fs
-     *         the filesystem that $path lives on
-     * @param  string|LocalFileInfo $path
-     *         the path that we want to convert
+     * @param  Filesystem $fs
+     *         the filesystem we are working with
+     * @param  string|PathInfo $path
+     *         what are we reading?
      * @param  OnFatal $onFatal
-     *         what do we do if we cannot do the type conversion?
-     * @return LocalFileInfo
+     *         what do we do when the reading fails?
+     * @return string
+     *         the full contents of $path
      */
-    public static function from(LocalFilesystem $fs, $path, OnFatal $onFatal) : LocalFileInfo
+    public static function using(LocalFilesystem $fs, $path, OnFatal $onFatal)
     {
-        // does it need converting?
-        if ($path instanceof LocalFileInfo) {
-            // no, it does not
-            return $path;
+        // what are we looking at?
+        $pathInfo = Internal\GetMetadataPathInfo::for($path);
+
+        if (!file_exists($pathInfo->getFullPath())) {
+            return [];
+        }
+        if (!Checks\IsFile::check($fs, $pathInfo)) {
+            throw $onFatal("not a file");
         }
 
-        // it needs converting
-        return Internal\GetFileInfo::for($fs, $path, $onFatal);
+        // TODO - add robustness
+        return json_decode(file_get_contents($pathInfo->getFullPath()), true);
     }
 }