+++ src/V1/Operations/PutFileMetadata.php
@@ -33,44 +33,62 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * @category  Libraries
- * @package   LocalFilesystem\V1\TypeConverters
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://ganbarodigital.github.io/php-mv-filesystem-plugin-local
  */
 
-namespace GanbaroDigital\LocalFilesystem\V1\TypeConverters;
+namespace GanbaroDigital\LocalFilesystem\V1\Operations;
 
 use GanbaroDigital\AdaptersAndPlugins\V1\PluginTypes\PluginClass;
+use GanbaroDigital\Filesystem\V1\Checks;
 use GanbaroDigital\LocalFilesystem\V1\Internal;
-use GanbaroDigital\LocalFilesystem\V1\LocalFileInfo;
 use GanbaroDigital\LocalFilesystem\V1\LocalFilesystem;
 use GanbaroDigital\MissingBits\ErrorResponders\OnFatal;
 
-class ToFileInfo implements PluginClass
+/**
+ * save metadata about a file
+ */
+class PutFileMetadata implements PluginClass
 {
     /**
-     * convert a path into a LocalFileInfo value
+     * save metadata about a file
      *
      * @param  LocalFilesystem $fs
-     *         the filesystem that $path lives on
-     * @param  string|LocalFileInfo $path
-     *         the path that we want to convert
+     *         the filesystem we are working with
+     * @param  string|PathInfo $path
+     *         what are we writing to?
+     * @param  array $metadata
+     *         what metadata do we want to attach to the file?
      * @param  OnFatal $onFatal
-     *         what do we do if we cannot do the type conversion?
-     * @return LocalFileInfo
+     *         what do we do when the write fails?
+     * @return void
      */
-    public static function from(LocalFilesystem $fs, $path, OnFatal $onFatal) : LocalFileInfo
+    public static function using(LocalFilesystem $fs, $path, array $metadata, OnFatal $onFatal)
     {
-        // does it need converting?
-        if ($path instanceof LocalFileInfo) {
-            // no, it does not
-            return $path;
+        // what are we looking at?
+        $pathInfo = Internal\GetMetadataPathInfo::for($path);
+
+        if (Checks\IsFolder::check($fs, $pathInfo)) {
+            throw $onFatal("already exists, is a folder");
         }
 
-        // it needs converting
-        return Internal\GetFileInfo::for($fs, $path, $onFatal);
+        $contents = json_encode($metadata);
+
+        // quick check - do we already have this content?
+        //
+        // we may be writing to a very slow / oversubscribed device,
+        // such as NFS.
+        // we may be writing to a replicating folder, managed by Dropbox
+        // or equivalent.
+        //
+        // we want to avoid writes where we can
+        if (Internal\IsDuplicateContent::check($pathInfo->getFullPath(), $contents)) {
+            return;
+        }
+
+        // TODO: add robustness
+        file_put_contents($pathInfo->getFullPath(), $contents);
     }
 }