+++ docs/changelog.md
@@ -22,3 +22,6 @@
   - added `GetFileContents` operation
   - added `PutFileContents` operation
   - added `Unlink` operation
+* Added support for key/value pair metadata
+  - added `GetFileMetadata` operation
+  - added `PutFileMetadata` operation